requirements.txt
altair==5.2.0
appdirs==1.4.4
blinker==1.7.0
cachetools==5.3.3
click==8.1.7
frozendict==2.4.0
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.42
html5lib==1.1
huggingface-hub==0.22.1
lxml==5.1.0
mpmath==1.3.0
multitasking==0.0.11
networkx==3.2.1
numpy==1.26.4
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.1
peewee==3.17.1
pillow==10.2.0
plotly==5.20.0
protobuf==4.25.3
pyarrow==15.0.2
pydeck==0.8.1b0
PySocks==1.7.1
python-dotenv==1.0.1
pytz==2024.1
regex==2023.12.25
safetensors==0.4.2
scipy==1.12.0
selenium==4.19.0
smmap==5.0.1
sortedcontainers==2.4.0
streamlit==1.32.2
sympy==1.12
tenacity==8.2.3
tokenizers==0.15.2
toml==0.10.2
toolz==0.12.1
torch==2.2.2+cpu
torchaudio==2.2.2+cpu
torchvision==0.17.2+cpu
tqdm==4.66.2
transformers==4.39.1
trio==0.25.0
trio-websocket==0.11.1
tzdata==2024.1
watchdog==4.0.0
wsproto==1.2.0
yfinance==0.2.37
